ons-retail-sales-rap/requirements.txt
pandas>=2.2
requests>=2.31
PyYAML>=6.0
matplotlib>=3.8
pytest>=8.0
jupyter>=1.1
